+++ src/govagent/hitl/__init__.py
@@ -2,10 +2,12 @@
 
 from .manager import HITLManager, ApprovalRequest, ApprovalStatus
 from .adapters import CLIAdapter, HITLAdapter
+from .slack_adapter import SlackJudiciaryAdapter
 
 # This allows: from govagent.hitl import HITLManager
 __all__ = [
     "HITLManager", 
+    "SlackJudiciaryAdapter",
     "ApprovalRequest", 
     "ApprovalStatus", 
     "CLIAdapter",